+++ src/hooks/Queries.js
@@ -1,8 +1,7 @@
 
-import {GET_USER_QUERY,   GET_USERS_QUERY, GET_USER_PROJECTS_QUERY, GET_ALL_USER_PROJECTS_QUERY, GET_ME_QUERY, GET_DEVPROJECTS_QUERY, GET_ONE_DEVPROJECT_QUERY  } from "./queries";
-import { LOGIN_MUTATION } from "./mutations";
+import {GET_USER_QUERY,   GET_USERS_QUERY, GET_USER_PROJECTS_QUERY, GET_ALL_USER_PROJECTS_QUERY, GET_ME_QUERY, GET_DEVPROJECTS_QUERY, GET_ONE_DEVPROJECT_QUERY  } from "../graphql/queries";
 
-import { useMutation, useQuery } from "@apollo/client";
+import { useQuery } from "@apollo/client";
 
 export function useGetUserQuery(id) {
  const { data, loading, error } = useQuery(GET_USER_QUERY, {
@@ -76,8 +75,4 @@ export function useGetOneDevProjectQuery(id) {
 }
 }
 
-export function useLoginMutation() {
-  const [login, { loading, error }] = useMutation(LOGIN_MUTATION);
-  return { login, loading, error };
-}
